refactor(action): move offline credential gating to its own module
Extract the analyzer credential drop into action_environment.py with a
single call site before mode dispatch, and move its tests to a dedicated
file, so the oversized action_runner.py grows only by the wiring. Bump
the two affected code-quality baseline entries accordingly.
