+++ ci/code-quality-baseline.json
@@ -184,7 +184,7 @@
     "scripts/ci/finish_extension_authority_stability.py::_move_authority_tests": 148,
     "scripts/codex-auto-resume-smoke.py::_queue_pending_request": 103,
     "scripts/stress_guard_daemon.py::run_stress": 133,
-    "src/codex_plugin_scanner/action_runner.py::main": 387,
+    "src/codex_plugin_scanner/action_runner.py::main": 388,
     "src/codex_plugin_scanner/checks/kimi.py::check_kimi_mcp_servers": 198,
     "src/codex_plugin_scanner/checks/marketplace.py::check_marketplace_json": 184,
     "src/codex_plugin_scanner/cli.py::_build_parser": 102,
@@ -470,7 +470,7 @@
     "scripts/ci/rust_authority_ownership_gate.py": 502,
     "scripts/mdm/macos/device-key-helper.swift": 501,
     "scripts/verify_release_registry.py": 523,
-    "src/codex_plugin_scanner/action_runner.py": 688,
+    "src/codex_plugin_scanner/action_runner.py": 690,
     "src/codex_plugin_scanner/checks/manifest.py": 501,
     "src/codex_plugin_scanner/checks/security.py": 924,
     "src/codex_plugin_scanner/guard/adapters/bounded_cli_hook_bridge.py": 758,
